worst case csv's

+++ QuickSortTester.java
@@ -6,9 +6,9 @@ public class QuickSortTester{
 	  for (int i = 1; i < 10000; i++) {
 	  int[] arr = QuickSort.buildArray(i,500);
 	  QuickSort.shuffle(arr);
-	  int startTime = System.nanoTime();
+	  double startTime = System.nanoTime();
 	  QuickSort.qsort(arr);
-	  int endTime = System.nanoTime();
+	  double endTime = System.nanoTime();
 	  System.out.println(endTime - startTime);
 	  }
 	  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~*/
@@ -17,22 +17,23 @@ public class QuickSortTester{
 	/*~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 	  for (int i = 1; i < 10000; i++) {
 	  int[] arr = QuickSort.buildArray(i,0);
-	  int startTime = System.nanoTime();
+	  double startTime = System.nanoTime();
 	  QuickSort.qsort(arr);
-	  int endTime = System.nanoTime();
+	  double endTime = System.nanoTime();
 	  System.out.println(endTime - startTime);
 	  }
 	  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~*/
 
 	//worst case 2: already sorted array
-	/*~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+	
 	  for (int i = 1; i < 10000; i++) {
 	  int[] arr = QuickSort.buildArrayIntegral(i);
-	  int startTime = System.nanoTime();
+	  double startTime = System.nanoTime();
 	  QuickSort.qsort(arr);
-	  int endTime = System.nanoTime();
+	  double endTime = System.nanoTime();
 	  System.out.println(endTime - startTime);
 	  }
+	  /*~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 	  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~*/
     }
 }
